requirements.txt
aiogram>=3.10.0
apscheduler>=3.10.4
selenium>=4.23.0
undetected-chromedriver>=3.5.5
python-dotenv>=1.0.1
